[CBRD-20149] XASL cache redesign (#125)
This revision covers two jira issues:
http://jira.cubrid.org/browse/CBRD-20149
http://jira.cubrid.org/browse/CBRD-20150

The XASL cache has been redesigned into a latch-free hash-table. XASL cache clones have been activated, they can be used if system parameter is set and they are protected by a mutex per XASL cache entry. All XASL cache code is now in separate files: xasl_cache.c/filter_pred_cache.c. Old code in query_executor.c has been removed.

* XASL cache structure:
 * Replaced the three hash tables based on query string, xasl ID and OID's with one lock-free hash table. The key is a modified XASL_ID which include SHA-1 generated from query string.
 * XASL_ID also includes a new field cache_flag. The flag is only modified using atomic CAS operations and is used to defer entry invalidation as long as it is still in use. cache_flag has two components: fix count and flags to mark entries for removal or for recompilation.
 * First_vpid/temp_vfid have been removed from XASL_ID. The XASL stream is cached in memory (in XASL_CACHE_ENTRY).
 * Inserting into cache: find or insert is called. If entry exists and if query must be recompiled, the existing entry is marked "to be recompiled" and another entry is inserted. Recompile note: the new inserted entry "inherits" time_stored from old existing entry; this prevents unnecessary entry invalidation when classes/serial suffer no changes.
 * Finding in hash by sha1 (called on prepare/execute): return entry found by sha1 generated from query hash string. The code that locks classes is included here.
 * Finding in hash by xasl_id (called on execute): time_stored must also match, to make sure the entry is still valid for its caller.
 * Invalidate entries after schema change: since oid-based hash table has been removed completely, invalidating entries based on OID must check all hash-table entries. It might be somewhat slower than before, but schema changes are rare and the hash-table is not blocked in the process.
 * Cleanup: is triggered when the maximum capacity of hash-table is reached. The hash-table is not blocked and new entries can be added. The cleaner will choose least recently used entries in the hash and will try to remove them.
 * Recompilation threshold check: RT check calls are limited to once per 10 minutes (can be changed). Max threshold is set to 10k pages (when tables grow above this value, the check is disabled). RT check is required if the cardinal has changed by a factor of 10x or 0.1x.
 * Removed "pinned" XASL implementation from server; no longer needed with the new implementation.

* XASL cached clones:
 * XASL clones can be cached if activate by user using max_plan_cache_clones system parameter.
 * One stack per entry is used to store clones. The stack starts with the size one and is iteratively doubled up to the maximum size when required.
 * Cached clones use global heap instead of local heap.
 * The cached clones stack is protected by mutex. Whenever a clone is claimed or retired, the mutex must be acquired first.

* Filter predicate cache:
 * Separated implementation from XASL cache.
 * Latch-free hash-table with mutexes on entries.
 * Clones activated by default.
 * Find in cache: find entry in hash table and claim a cached filter predicate expression.
 * Retire to cache: add filter predicate expression to an entry in cache; if entry does not exist, it is inserted; if entry is full, filter predicate expression is freed.
 * Invalidate entries after schema change: very similar to xasl cache.
 * Cleanup: similar to xasl cache.

* Lock-free hash table:
 * Using mutex flags merged into one flag only: uses mutex.
 * Allow lock-free transaction "promotion" from transaction ID not incremented to transaction ID incremented.
 * Modify return value of function that cannot fail to void.
 * Added transaction systems for XASL cache and filter predicate cache.
 * Merged lf_list_insert and lf_list_find_or_insert into lf_list_insert_internal function.
 * Added function lf_hash_insert_given; lf_hash_insert, lf_hash_find_or_insert and lf_hash_insert_given output if entry was inserted.
 * Added function lf_hash_delete_already_locked.
 * Added lf_tran_start_with_mb and lf_tran_end_with_mb
 * Hash iterate reset transaction for each bucket.

Other changes:
* Fixed a series of warnings.
* rewrite session_remove_expired_sessions to avoid deletes during iterate.
* include user in query hash string
* updated sha1 to be used for xasl cache.
* updated binary_heap to be used for xasl cache and filter predicate cache cleanup.

TODO in the future:
* Extend RT checks to executing prepared statements.
* Analyze and rewrite existing pinned XASL implementation (might be needed to force recompile after RT checks).
* Additional cleanup criterias: periodic, if total memory exceeds a threshold.
* Match xasl cache entry with XASL_ID using a different criteria than time stored; XASL entry should be invalidated only by schema/serial changes.
* Modify XASL stream: remove xasl_id. store xasl_header not as pointer.
* Handle MRO in query hash string.
* Drop all queries
* Remove "pinned" XASL routine from client/driver.

+++ src/connection/connection_sr.c
@@ -2451,7 +2451,7 @@ css_return_queued_data_timeout (CSS_CONN_ENTRY * conn, unsigned short rid, char 
 		  int r;
 		  struct timespec abstime;
 
-		  abstime.tv_sec = time (NULL) + waitsec;
+		  abstime.tv_sec = (int) time (NULL) + waitsec;
 		  abstime.tv_nsec = 0;
 
 		  r = thread_suspend_timeout_wakeup_and_unlock_entry (thrd, &abstime, THREAD_CSS_QUEUE_SUSPENDED);
